requirements.txt
numpy<2
scikit-surprise
pandas
lightgbm
scikit-learn
pymongo
python-dotenv
Faker